# api/openapi/control/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Control API
  version: 1.0.0
  description: V1 HTTP contract for Identity, platform administration, Tenant membership, Agent authoring, Runtime Profile authoring, and Deployment publication.
servers:
  - url: /
...
paths:
  /v1/tenants/{tenant_id}/runs:
    parameters:
      - $ref: '#/components/parameters/TenantID'
    get:
      operationId: listRuns
      tags: [Operations]
      summary: List tenant Run records
      description: Reads the Worker-owned execution ledger through Control; never returns message bodies, credentials, or Session content.
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Run page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RunPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {description: Runtime observation source unavailable}
  /v1/tenants/{tenant_id}/runs/{run_id}:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: run_id
        in: path
        required: true
        schema: {type: string}
    get:
      operationId: getRun
      tags: [Operations]
      summary: Get a Run execution timeline
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: Run detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RunDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Run not found in this tenant}
        '503': {description: Runtime observation source unavailable}
  /v1/tenants/{tenant_id}/audit-events:
    parameters:
      - $ref: '#/components/parameters/TenantID'
    get:
      operationId: listAuditEvents
      tags: [Operations]
      summary: List tenant business audit events
      description: Merges Control-owned configuration facts and Worker-owned execution facts after tenant authorization.
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Audit page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {description: An audit source is unavailable}
  /v1/tenants/{tenant_id}/usage-policy:
    parameters:
      - $ref: '#/components/parameters/TenantID'
    get:
      operationId: getTenantUsagePolicy
      tags: [Operations]
      summary: Read the current tenant usage policy
      description: Active members may read the current policy. Absence is represented as a disabled revision-zero policy.
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: Current policy
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UsagePolicy'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
    put:
      operationId: replaceTenantUsagePolicy
      tags: [Operations]
      summary: Replace the current tenant usage policy
      description: Owner-only, CAS-protected replacement. Revision is concurrency control; V1 does not expose policy history.
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/UsagePolicyIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReplaceUsagePolicyRequest'}
      responses:
        '200':
          description: Replaced policy
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UsagePolicy'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/tenants/{tenant_id}/usage-summary:
    parameters:
      - $ref: '#/components/parameters/TenantID'
    get:
      operationId: getTenantUsageSummary
      tags: [Operations]
      summary: Read Worker-owned usage accounting for the latest period
      description: Unknown Provider usage retains its reservation and is never reported as zero. Estimated cost is not a Provider invoice.
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: Current usage projection
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UsageSummary'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/tenants/{tenant_id}/tool-approvals:
    parameters:
      - $ref: '#/components/parameters/TenantID'
    get:
      operationId: listToolApprovals
      tags: [Operations]
      summary: List exact side-effecting tool operations awaiting or carrying a decision
      description: Active tenant members may read redacted operation summaries. Raw arguments and provider results are not returned.
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Tool approval page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ToolApprovalPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {description: Worker approval source unavailable}
  /v1/tenants/{tenant_id}/tool-approvals/{operation_id}/decision:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: operation_id
        in: path
        required: true
        schema: {type: string, pattern: '^tap_[0-9a-f]{64}$'}
    post:
      operationId: decideToolApproval
      tags: [Operations]
      summary: Approve or reject one exact test-ticket status operation
      description: OWNER-only. expected_arguments_digest prevents a stale page or changed parameters from consuming an approval. Replaying the same decision is idempotent.
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ToolApprovalDecisionRequest'}
      responses:
        '200':
          description: Durable decision or identical replay
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ToolApprovalDecisionResponse'}
        '400': {description: Invalid decision}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Approval not found in this tenant}
        '409': {description: "Stale digest, opposite decision, expiry, or execution already started"}
        '503': {description: Worker approval source unavailable}
...
  /v1/auth/login:
    post:
      operationId: loginWithPassword
      summary: Log in with a local username and password
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication succeeded and a server-side session was created.
          headers:
            Set-Cookie:
              description: HttpOnly opaque session cookie. The token is never returned in JSON.
              schema: {type: string}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LoginResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429':
          description: The per-process, per-peer login attempt limit was exceeded.
          headers:
            Retry-After:
              description: Seconds until the current fixed window expires.
              schema: {type: integer}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: control_session
  parameters:
    TenantID:
      name: tenant_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
    UsagePolicyIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable key for one logical policy replacement; identical retries return the original result and changed payloads return 409.
      schema: {type: string, minLength: 8, maxLength: 128, pattern: '^[!-~]+$'}
...
    Offset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    BadRequest:
      description: Request input is invalid.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Unauthorized:
      description: A valid session is required.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Forbidden:
      description: The authenticated user is not authorized or must change a temporary password.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    Conflict:
      description: The requested state change conflicts with current state.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    ServiceUnavailable:
      description: A required owner service or credential-metadata dependency is temporarily unavailable.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    InternalError:
      description: The request could not be completed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
    UsagePolicy:
      type: object
      additionalProperties: false
      required: [schema_version, tenant_id, revision, enabled, im, requests, execution, tokens]
      properties:
        schema_version: {type: integer, const: 1}
        tenant_id: {type: string, minLength: 1, maxLength: 128}
        revision: {type: integer, minimum: 0, maximum: 9007199254740991}
        enabled: {type: boolean}
        im:
          type: object
          additionalProperties: false
          required: [allow_all, rules]
          properties:
            allow_all: {type: boolean}
            rules:
              type: array
              maxItems: 256
              items:
                type: object
                additionalProperties: false
                required: [account_id, user_ids, group_ids]
                properties:
                  account_id: {type: string, minLength: 1, maxLength: 128}
                  binding_id: {type: string, minLength: 1, maxLength: 128}
                  user_ids: {type: array, maxItems: 2048, items: {type: string, minLength: 1, maxLength: 256}}
                  group_ids: {type: array, maxItems: 2048, items: {type: string, minLength: 1, maxLength: 256}}
        requests:
          type: object
          additionalProperties: false
          required: [tenant_per_minute, user_per_minute]
          properties:
            tenant_per_minute: {type: integer, minimum: 0, maximum: 1000000}
            user_per_minute: {type: integer, minimum: 0, maximum: 1000000}
        execution:
          type: object
          additionalProperties: false
          required: [max_concurrent_runs]
          properties:
            max_concurrent_runs: {type: integer, minimum: 0, maximum: 100000}
        tokens:
          type: object
          additionalProperties: false
          required: [period_seconds, limit, reservation_per_run, input_micros_per_million_tokens, output_micros_per_million_tokens]
          properties:
            period_seconds: {type: integer, minimum: 0, maximum: 31536000}
            limit: {type: integer, minimum: 0, maximum: 9007199254740991}
            reservation_per_run: {type: integer, minimum: 0, maximum: 9007199254740991}
            input_micros_per_million_tokens: {type: integer, minimum: 0, maximum: 1000000000}
            output_micros_per_million_tokens: {type: integer, minimum: 0, maximum: 1000000000}
    ReplaceUsagePolicyRequest:
      type: object
      additionalProperties: false
      required: [expected_revision, policy]
      properties:
        expected_revision: {type: integer, minimum: 0, maximum: 9007199254740991}
        policy: {$ref: '#/components/schemas/UsagePolicy'}
    UsageSummary:
      type: object
      additionalProperties: false
      required: [tenant_id, policy_revision, period_seconds, token_limit, used_tokens, reserved_tokens, unknown_usage_count, pending_usage_count, estimated_cost_micros]
      properties:
        tenant_id: {type: string, minLength: 1, maxLength: 128}
        policy_revision: {type: integer, minimum: 0, maximum: 9007199254740991}
        period_start: {type: string, format: date-time}
        period_seconds: {type: integer, minimum: 0, maximum: 31536000}
        token_limit: {type: integer, minimum: 0, maximum: 9007199254740991}
        used_tokens: {type: integer, minimum: 0, maximum: 9007199254740991}
        reserved_tokens: {type: integer, minimum: 0, maximum: 9007199254740991}
        unknown_usage_count: {type: integer, minimum: 0, maximum: 9007199254740991}
        pending_usage_count: {type: integer, minimum: 0, maximum: 9007199254740991}
        estimated_cost_micros: {type: integer, minimum: 0, maximum: 9007199254740991}
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 256}
        password: {type: string, minLength: 1, maxLength: 1024, format: password}
    LoginResponse:
      type: object
      additionalProperties: false
      required: [user, password_change_required]
      properties:
        user: {$ref: '#/components/schemas/CurrentUser'}
        password_change_required: {type: boolean}
    CurrentUser:
      type: object
      additionalProperties: false
      required: [id, username, display_name]
      properties:
        id: {type: string}
        username: {type: string}
        display_name: {type: string}
...
    RunSummary:
      type: object
      required: [run_id, session_id, status, stage, attempts, usage_status, input_tokens, output_tokens, total_tokens, reply_status, accepted_at]
      properties:
        run_id: {type: string}
        session_id: {type: string}
        status: {type: string, enum: [QUEUED, RUNNING, RETRY_WAIT, SUCCEEDED, FAILED]}
        stage: {type: string, enum: [QUEUED, EXECUTION, MEMORY, REPLY_PUBLISH, REPLY_HANDOFF, COMPLETED, FAILED]}
        wait_reason: {type: string}
        failure_reason: {type: string}
        attempts: {type: integer, minimum: 0}
        usage_status: {type: string, enum: [UNAVAILABLE, PARTIAL, COMPLETE]}
        input_tokens: {type: integer, minimum: 0}
        output_tokens: {type: integer, minimum: 0}
        total_tokens: {type: integer, minimum: 0}
        memory_status: {type: string}
        reply_status: {type: string, enum: [NOT_CREATED, PENDING, HANDED_OFF]}
        accepted_at: {type: string, format: date-time}
        execution_deadline: {type: string, format: date-time}
    RunAttempt:
      type: object
      additionalProperties: false
      required: [attempt_id, worker_id, generation, status, created_at]
      properties:
        attempt_id: {type: string}
        worker_id: {type: string}
        generation: {type: integer, minimum: 1}
        status: {type: string}
        reason: {type: string}
        created_at: {type: string, format: date-time}
        started_at: {type: string, format: date-time}
        ended_at: {type: string, format: date-time}
    TimelineEvent:
      type: object
      additionalProperties: false
      required: [source, category, status, occurred_at]
      properties:
        source: {type: string, enum: [worker, gateway, control]}
        category: {type: string}
        status: {type: string}
        reason: {type: string}
        occurred_at: {type: string, format: date-time}
        attributes: {type: object, additionalProperties: true}
    RunPage:
      type: object
      additionalProperties: false
      required: [runs, offset, limit, total]
      properties:
        runs: {type: array, items: {$ref: '#/components/schemas/RunSummary'}}
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
    RunDetail:
      allOf:
        - $ref: '#/components/schemas/RunSummary'
        - type: object
          required: [admission_id, attempt_log, timeline, coverage]
          properties:
            admission_id: {type: string}
            manifest_ref: {type: string}
            manifest_digest: {type: string}
            session_head: {type: string}
            attempt_log: {type: array, items: {$ref: '#/components/schemas/RunAttempt'}}
            timeline: {type: array, items: {$ref: '#/components/schemas/TimelineEvent'}}
            coverage: {type: array, items: {type: string}}
    AuditEvent:
      type: object
      additionalProperties: false
      required: [event_id, source, category, action, outcome, resource_type, resource_id, occurred_at]
      properties:
        event_id: {type: string}
        source: {type: string, enum: [control, worker, gateway]}
        category: {type: string}
        action: {type: string}
        outcome: {type: string}
        actor_id: {type: string}
        resource_type: {type: string}
        resource_id: {type: string}
        reason: {type: string}
        occurred_at: {type: string, format: date-time}
        attributes: {type: object, additionalProperties: true}
    AuditPage:
      type: object
      additionalProperties: false
      required: [events, offset, limit, total]
      properties:
        events: {type: array, items: {$ref: '#/components/schemas/AuditEvent'}}
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
    ToolApproval:
      type: object
      additionalProperties: false
      required: [operation_id, tenant_id, run_id, attempt_id, node_id, tool_name, tool_resource, capability, target, parameter_summary, arguments_digest, status, requested_at, expires_at]
      properties:
        operation_id: {type: string, pattern: '^tap_[0-9a-f]{64}$'}
        tenant_id: {type: string, minLength: 1}
        run_id: {type: string, minLength: 1}
        attempt_id: {type: string, minLength: 1}
        node_id: {type: string, minLength: 1}
        tool_name: {type: string, minLength: 1}
        tool_resource: {type: string, minLength: 1}
        capability: {type: string, enum: [test.ticket.status.update]}
        target: {type: string, minLength: 1}
        parameter_summary: {type: string, minLength: 1}
        arguments_digest: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
        status: {type: string, enum: [PENDING, APPROVED, REJECTED, EXPIRED, EXECUTING, SUCCEEDED, UNKNOWN]}
        requested_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        decided_by: {type: string}
        decided_at: {type: string, format: date-time}
        decision_reason: {type: string}
        execution_started_at: {type: string, format: date-time}
        execution_finished_at: {type: string, format: date-time}
        result_summary: {type: string}
        result_digest: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
    ToolApprovalPage:
      type: object
      additionalProperties: false
      required: [operations, offset, limit, total]
      properties:
        operations: {type: array, items: {$ref: '#/components/schemas/ToolApproval'}}
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
    ToolApprovalDecisionRequest:
      type: object
      additionalProperties: false
      required: [action, expected_arguments_digest]
      properties:
        action: {type: string, enum: [approve, reject]}
        reason: {type: string, maxLength: 500}
        expected_arguments_digest: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
    ToolApprovalDecisionResponse:
      type: object
      additionalProperties: false
      required: [operation, outcome]
      properties:
        operation: {$ref: '#/components/schemas/ToolApproval'}
        outcome: {type: string, enum: [DECIDED, REPLAY]}
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}